library-manager/requirements.txt
appdirs==1.4.4
asgiref==3.3.0
autopep8==1.4.4
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2019.3.9
chardet==3.0.4
comtypes==1.1.7
cssselect==1.1.0
decouple==0.0.7
defusedxml==0.6.0
dj-database-url==0.5.0
django-filter==2.4.0
django-heroku==0.3.1
fake-useragent==0.1.11
gunicorn==20.0.4
html5lib==1.1
idna==2.10
inflect==4.1.0
jaraco.itertools==5.0.0
lxml==4.6.1
more-itertools==8.6.0
MouseInfo==0.1.3
parse==1.18.0
Pillow==8.0.1
plyer==1.4.3
psutil @ file:///C:/Users/Rkct/Downloads/psutil-5.7.3-cp39-cp39-win_amd64.whl
psycopg2==2.8.6
PyAudio @ file:///C:/Users/Rkct/Downloads/PyAudio-0.2.11-cp39-cp39-win_amd64.whl
PyAutoGUI==0.9.52
pycodestyle==2.6.0
pyee==7.0.4
PyGetWindow==0.0.9
pyjokes==0.6.0
PyMsgBox==1.0.9
pyperclip==1.8.1
pypiwin32==223
pyppeteer==0.2.2
pyquery==1.4.1
PyRect==0.1.4
PyScreeze==0.1.26
pyttsx3==2.90
PyTweening==1.0.3
pytz==2020.4
pywin32==228
requests==2.24.0
requests-html==0.10.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
SpeechRecognition==3.8.1
sqlparse==0.4.1
tqdm==4.51.0
urllib3==1.25.11
w3lib==1.22.0
webencodings==0.5.1
websockets==8.1
whitenoise==5.2.0
wikipedia==1.4.0
wolframalpha==4.0.0
xmltodict==0.12.0
